Split the difference renderer into `chiaroscuro.render`
`chiaroscuro.core` needed escapade, escritoire, dendrology and iridescence for
one thing: rendering a `Juxtaposition` as a table of differences. Structural
contrast itself — Contrastable, Decomposable, Decomposition, Similarity —
needs none of it, and is what most consumers actually want.

The renderer and its palette move to a new `chiaroscuro.render`. A companion
object must share a file with its class, so the members leave `Juxtaposition`'s
companion and become toplevel; the given is named `juxtapositionTeletype` so it
can be imported and exported.

That takes it out of implicit scope, and escapade renders any `Showable` value
unstyled, so probably's `cmp.teletype` silently resolved to the fallback and
printed `Collation(Text,List(…))` instead of the difference table. Verified by
disabling the given and watching probably still compile. `AnsiRenderer` — the
only site in the tree that renders a `Juxtaposition` — now imports it
explicitly, so removing the given fails the build rather than degrading the
output, and two tests pin the rendered form of an equal and a differing pair.
